降低统一位置头拟合的内存峰值
one-hot 编码直接生成 float32，避免先分配约 8.76 GiB 的 float64 临时矩阵。使用冻结校准集重放后，summary、操作身份、最终 Top 与位置分数文件 SHA-256 均与原实现完全一致。

+++ scripts/fit-enriched-hierarchical-adjudicator.py
@@ -50,7 +50,7 @@ def encode_combined(
     values = pd.get_dummies(
         combined[columns],
         columns=[column for column in columns if combined[column].dtype == object],
-        dtype=float,
+        dtype=np.float32,
     ).replace([np.inf, -np.inf], np.nan).fillna(0).astype(np.float32)
     return values, len(development), list(values.columns)
